pyproject.toml
[project]
name = "to-do-list"
version = "0.1.0"
description = ""
authors = [
    {name = "Narmina",email = "29mehtiyevanarmina08@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "fastapi (>=0.116.1,<0.117.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.9,<1.0.0)"
]

[tool.poetry]
packages = [{include = "to_do_list", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
